final/requirements.txt
openai==0.28.1
python-dotenv
flask
flask_cors
requests
googlemaps